refactor: Improve folder and files structure and fix some material-table console warnings
